+++ doc/changelog.md
@@ -1,12 +1,15 @@
 # ChangeLog
 
 * **0.11.5-M1**
+  - Add a `user` parameter to the assembly configuration so that the added files are created for this user 
+    (#53, #175) 
   - Fix problem when creating intermediate archive for collecting assembly files introduced with #139. The 
     container can be now set with "mode" in the assembly configuration with the possible values `dir`, `tar`, `tgz`
     and `zip` (#171)
   - Workaround Docker problem when using an implicit registry `index.docker.io` when no registry is explicitly given. 
   - Fixed references to docker hub in documentation (#169)
   - Fixed registry authentication lookup (#146)
+  
   
 * **0.11.4**
   - Fixed documentation for available properties